🎨 Break down _layout into layout + effects files
