+++ src/components/canvas/_effects.js
@@ -1,0 +1,58 @@
+import { useFrame, useThree } from 'react-three-fiber'
+import { HTML } from '@react-three/drei'
+import * as THREE from 'three'
+
+import { ContactShadows } from '@react-three/drei'
+import { EffectComposer, Noise, Vignette } from '@react-three/postprocessing'
+
+import { defaultConfig } from '@/helpers/particles/config'
+
+import ParticleField from 'react-particles-webgl'
+
+const ParticlesBackground = () => (
+  <div id='particle-field' className='relative w-screen h-screen'>
+    <ParticleField config={defaultConfig} />
+  </div>
+)
+
+const Rig = () => {
+  const { camera, mouse } = useThree()
+  const vec = new THREE.Vector3()
+  return useFrame(() =>
+    camera.position.lerp(
+      vec.set(mouse.x * 2, mouse.y * 1, camera.position.z),
+      0.02
+    )
+  )
+}
+
+const Effects = () => {
+  return (
+    <>
+      <HTML prepend fullscreen zIndexRange={[100, 0]}>
+        <ParticlesBackground />
+      </HTML>
+
+      <fog attach='fog' args={[0xf9fafb, 60, 70]} />
+
+      <ambientLight color={new THREE.Color(0xf9fafb)} intensity={0.5} />
+
+      <ContactShadows
+        rotation={[Math.PI / 2, 0, 0]}
+        position={[0, -8, 0]}
+        opacity={0.75}
+        width={140}
+        height={140}
+        blur={1}
+        far={9}
+      />
+      <EffectComposer>
+        <Noise opacity={0.02} />
+        <Vignette eskil={false} offset={0.1} darkness={0.4} />
+      </EffectComposer>
+      <Rig />
+    </>
+  )
+}
+
+export default Effects
